reorganise frontend assets + update webpack config
