+++ package.json
@@ -5,13 +5,13 @@
     "scripts": {
         "start:dev": "nodemon -L --inspect=0.0.0.0:9229 -e .css,.js,.json,.njk --ignore public/dist/ --exec npm run buildRun:dev",
         "start": "node ./bin/www",
-        "sass": "node-sass src/sass/all.scss public/stylesheets/all.css && node-sass src/sass/all-ie8.scss public/stylesheets/all-ie8.css",
+        "sass": "node-sass src/sass/all.scss public/dist/css/all.css && node-sass src/sass/all-ie8.scss public/dist/css/all-ie8.css",
         "test": "npm run test:node && npm run test:dom",
         "test:dom": "jest --config=jest.config.jsdom.js",
         "test:node": "jest --config=jest.config.js",
         "coverage": "jest --coverage",
-        "build:dev": "npm run copy:css && npm run copy:js && webpack --config webpack.config.js --mode development",
-        "build": "npm run copy:css && npm run copy:js && webpack --config webpack.config.prod.js --mode production",
+        "build:dev": "webpack --config webpack.config.js --mode development",
+        "build": "webpack --config webpack.config.prod.js --mode production",
         "lint": "eslint .",
         "lint:fix": "eslint --fix .",
         "buildRun:dev": "npm run build:dev && node ./bin/www",
