+++ page/page.njk
@@ -3,21 +3,19 @@
 {% block pageTitle %}Claim criminal injuries compensation - GOV.UK{% endblock %}
 {% block head %}
   <!--[if !IE 8]><!-->
-    <link href="/govuk-frontend/all.css" rel="stylesheet" />
+    <link href="/dist/css/all.css" rel="stylesheet" />
   <!--<![endif]-->
 
   {# For Internet Explorer 8, you need to compile specific stylesheet #}
   {# see https://github.com/alphagov/govuk-frontend/blob/master/docs/installation/supporting-internet-explorer-8.md #}
   <!--[if IE 8]>
-    <link href="/govuk-frontend/all-ie8.css" rel="stylesheet" />
+    <link href="/dist/css/all-ie8.css" rel="stylesheet" />
   <![endif]-->
 
   {# For older browsers to allow them to recognise HTML5 elements such as `<header>` #}
   <!--[if lt IE 9]>
     <script src="/html5-shiv/html5shiv.js"></script>
   <![endif]-->
-  <link rel="stylesheet" href="/dist/css/accessible-autocomplete.css" />
-  <link rel="stylesheet" href="/dist/css/accessible-autocomplete-wrapper.css" />
 {% endblock %}
 
 
@@ -85,9 +83,7 @@
             SESSION_DURATION: {{ CW_SESSION_DURATION }}
         };
     </script>
-    <script src="/govuk-frontend/all.js"></script>
     <script nonce="{{ nonce }}">window.GOVUKFrontend.initAll()</script>
-    <script src="/dist/js/autocomplete.min.js"></script>
     <script src="/dist/js/bundle.js"></script>
 {% endblock %}
