+++ src/sass/all.scss
@@ -1,3 +1,5 @@
 @import "node_modules/govuk-frontend/govuk/all";
 @import "src/modules/cookie-banner/cookie-banner";
-@import "components/cica/modal/modal"
+@import "components/cica/modal/modal";
+@import "src/modules/autocomplete/accessible-autocomplete.scss";
+@import "src/modules/autocomplete/accessible-autocomplete-wrapper.scss";
